2.0.0 release
Updates module code to work with webpack. This new style removes
the need to initialise with window (e.g `require('please-ajax')(window)`)
which is a breaking change, hence 2.0.0, and removes compatibility
with node. I decided that was safe, since this modules biggest
feature is the form uploads that support IE9. Remind me to write
docs for that...